% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\name{add_entry_tbl}
\alias{add_entry_tbl}
\alias{modify_entry_tbl}
\alias{delete_entry_tbl}
\title{Add entry within specified table}
\usage{
add_entry_tbl(con = NULL, tbl = NULL, ...)

modify_entry_tbl(con = NULL, tbl = NULL, ...)

delete_entry_tbl(con = NULL, tbl = NULL, ...)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}

\item{...}{a vector of column names in the specified table}
}
\description{
Add entry within specified table
}
\section{Functions}{
\itemize{
\item \code{modify_entry_tbl()}: Modify entry within specified table.

\item \code{delete_entry_tbl()}: Delete entry within specified table.

}}
