% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\name{bulk_insert_entry}
\alias{bulk_insert_entry}
\title{Add entry within specified table}
\usage{
bulk_insert_entry(con = NULL, tbl = NULL, data = NULL)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}

\item{data}{a data.frame}
}
\description{
Add entry within specified table
}
