% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_insert_test}
\alias{bulk_insert_test}
\title{Table for bulk insert test}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
bulk_insert_test
}
\description{
A simple data.frame for bulk insert entry
}
\keyword{datasets}
