% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censored_data}
\alias{censored_data}
\title{Censored data exemple}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\usage{
censored_data
}
\description{
A simple data.frame which contains exemples of censored data
}
\keyword{datasets}
