% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_check_misc.R
\name{check_fields_exist}
\alias{check_fields_exist}
\title{Check if fields present in table}
\usage{
check_fields_exist(con, tbl, fields)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}

\item{fields}{a vector of column names in the specified table}
}
\description{
Check if fields present in table
}
\examples{
\dontrun{
     check_fields_exist(con, "species", species_id = "test")
}
}
