% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_check_misc.R
\name{check_fields_pkeys}
\alias{check_fields_pkeys}
\title{Check if pkeys presents}
\usage{
check_fields_pkeys(con, tbl, fields)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}

\item{fields}{a vector of column names in the specified table}
}
\description{
Check if pkeys presents
}
\examples{
\dontrun{
 check_fields_pkeys(con, "species", fields = c("species_id"))
}
}
