% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info_misc.R
\name{check_pkeys_fields}
\alias{check_pkeys_fields}
\title{Check if pkeys presents}
\usage{
check_pkeys_fields(con, tbl, fields)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}

\item{fields}{a vector of column names in the specified table}
}
\value{
Logical test result.
}
\description{
Check if pkeys presents
}
\examples{
\dontrun{
 check_pkeys_fields(con, "species", fields = c("species_id"))
}
}
