% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs_integration.R
\name{dbs_integration}
\alias{dbs_integration}
\title{Get integrated data}
\usage{
dbs_integration()
}
\value{
a data.frame of all contaminant measurements
}
\description{
This function is a quick integration procedure among all Excel files.
}
