% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{detect_cens}
\alias{detect_cens}
\title{Detect censored measurement}
\usage{
detect_cens(x)
}
\arguments{
\item{x}{lab measurements vector}
}
\description{
Censored data are detected by the presence of the symbol \link{<} .
}
