% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\name{add_entry_tbl}
\alias{add_entry_tbl}
\title{DB getters}
\usage{
add_entry_tbl(tbl, ...)
}
\arguments{
\item{tbl}{a character name of the table}

\item{...}{a vector of column names in the specified table}
}
\description{
Retrieve specific information from our admin database.
}
\section{Functions}{
\itemize{
\item \code{add_entry_tbl()}: Get species name list.

}}
