% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info_misc.R
\name{get_tbl_fields_notnull}
\alias{get_tbl_fields_notnull}
\title{Get table not null fields}
\usage{
get_tbl_fields_notnull(con, tbl)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}
}
\value{
Not null fields as a character vector.
}
\description{
Get table not null fields
}
\examples{
\dontrun{
 get_tbl_fields_notnull("species")
}
}
