% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info_misc.R
\name{get_tbl_fields_pkey}
\alias{get_tbl_fields_pkey}
\title{Get table primary key fields}
\usage{
get_tbl_fields_pkey(con, tbl)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}
}
\value{
Primary key fields as a character vector.
}
\description{
Get table primary key fields
}
\examples{
\dontrun{
 get_tbl_fields_pkey("species")
}
}
