% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info_misc.R
\name{get_tbl_info}
\alias{get_tbl_info}
\title{Get table fields properties (fkey, pkey, unique constraints etc.)}
\usage{
get_tbl_info(con = NULL, tbl = NULL)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}
}
\value{
data.frame with 6 columns:
\itemize{
\item cid: field position
\item name: field name
\item type: field type
\item notnull: is not null constraint apply to this field?
\item dflt_value: field default value
\item pk: is this field a primary key?
}
}
\description{
Get table fields properties (fkey, pkey, unique constraints etc.)
}
\examples{
\dontrun{
 get_tbl_info("species")
}
}
