% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\name{insert_entry}
\alias{insert_entry}
\alias{modify_entry}
\alias{delete_entry}
\title{Add entry within specified table}
\usage{
insert_entry(con = NULL, tbl = NULL, ...)

modify_entry(con = NULL, tbl = NULL, ...)

delete_entry(con = NULL, tbl = NULL, ...)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}

\item{...}{a vector of column names in the specified table}
}
\description{
Add entry within specified table
}
\section{Functions}{
\itemize{
\item \code{modify_entry()}: Modify entry within specified table.

\item \code{delete_entry()}: Delete entry within specified table.

}}
