% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgr_measurements.R
\name{itgr_measurements}
\alias{itgr_measurements}
\title{Measurements integration from scattered Excel files.}
\usage{
itgr_measurements()
}
\value{
a data.frame of all contaminant measurements
}
\description{
Function performing measurements integration procedure from scattered Excel files.
}
