% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itgr_samples_info.R
\name{itgr_samples_info}
\alias{itgr_samples_info}
\title{Samples informations integration from scattered Excel files}
\usage{
itgr_samples_info()
}
\value{
a data.frame of all sample information
}
\description{
Function performing sample informations integration from scattered Excel files.
}
