% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{remove_cens}
\alias{remove_cens}
\title{Remove censored measurement symbol (<)}
\usage{
remove_cens(x)
}
\arguments{
\item{x}{lab measurements vector}
}
\description{
Remove censored measurement symbol (<)
}
