% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_search.R
\name{search_tbl}
\alias{search_tbl}
\title{search_tbl}
\usage{
search_tbl(con = NULL, tbl = NULL, ..., nfetch = 100)
}
\arguments{
\item{con}{connexion object returned by DBI::dbConnect()}

\item{tbl}{a character name of the table}

\item{...}{table fields with search value}

\item{nfetch}{data fetch size (default = 100)}
}
\value{
A data.frame with query results
}
\description{
search_tbl
}
\examples{
\dontrun{
 search_tbl(con, "species", genus = "\%Poo\%")
}
}
