% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbl_info_test}
\alias{tbl_info_test}
\title{Table informations example}
\format{
An object of class \code{data.frame} with 3 rows and 6 columns.
}
\usage{
tbl_info_test
}
\description{
A simple data.frame which contains example of table returned by get_tbl_info for testing purpose
}
\keyword{datasets}
