% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_info_misc.R
\name{check_notnull_fields}
\alias{check_notnull_fields}
\title{Check if not null fields present}
\usage{
check_notnull_fields(tbl, fields)
}
\arguments{
\item{tbl}{a character name of the table}

\item{fields}{a vector of column names in the specified table}
}
\value{
Logical test result.
}
\description{
Check if not null fields present
}
\examples{
\dontrun{
 check_notnull_fields("species", fields = c("species_id"))
}

}
