% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_search.R
\name{search_tbl}
\alias{search_tbl}
\title{search_tbl}
\usage{
search_tbl(tbl, ...)
}
\arguments{
\item{tbl}{a character name of the table}

\item{...}{table fields with search value}
}
\value{
A data.frame with query results
}
\description{
search_tbl
}
\examples{
\dontrun{
 search_tbl("species", genus = "\%Poo\%")
}
}
