% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{uncensored}
\alias{uncensored}
\title{Detect censored data}
\usage{
uncensored(df = NULL, cols = c(NULL), keep_cens = TRUE)
}
\arguments{
\item{template}{a template. Use "?" to list available template.}

\item{data}{data passed to \code{\link[whisker:whisker.render]{whisker::whisker.render()}}.}

\item{...}{further arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
The return value, if any, from executing the function.
}
\description{
For each column specified, censored data are detected by the presence of the symbol \link{<} . This function returned a data.frame with two columns by
}
\examples{
\dontrun{
custom_render("report1",
  data = list(title = "templex"),
  envir = list(geom = 1)
)
}
}
