% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{uncensored}
\alias{uncensored}
\title{Manage censored data}
\usage{
uncensored(df = NULL, cols = c(NULL), keep_cens = TRUE)
}
\arguments{
\item{df}{a data.frame.}

\item{cols}{a vector of character.}

\item{keep_cens}{a boolean. If TRUE, assigns the logicical test result (presence of absence of symbol <) to a new column}
}
\value{
The return value, if any, from executing the function.
}
\description{
For each column specified, censored data are detected by the presence of the symbol \link{<} . This function removes all censored data in the specified columns. keep_cens=TRUE assigns the logicical test result (presence of absence of symbol <) to a new column, to keep track of the original censored data.
}
\examples{
\dontrun{
custom_render("report1",
  data = list(title = "templex"),
  envir = list(geom = 1)
)
}
}
